refactor(trusty-code): split main.rs below the 500 SLOC cap (#4434)
Why: crates/trusty-code/src/main.rs reached 498 of the mechanically-enforced
500-SLOC production cap (scripts/check_line_cap.sh, the line-cap CI workflow,
and the pre-commit hook) after #4424 added Command::Tui. With two lines of
headroom the NEXT change to the file — any change — fails the gate before it
starts, and more milestone work is queued behind it.

What: moves the --legacy-in-process run-task path out of main.rs into
crates/trusty-code/src/cli/legacy_run_task.rs, next to every other subcommand
handler. That block is the one place in main.rs that held real logic rather
than clap dispatch: the in-process wrapper (run_task, now
legacy_run_task::run) plus the two helpers only it uses
(validate_agent_name's path-traversal guard, build_llm_client's
DispatchingLlmClient construction), the TCODE_ENGINEER_MODEL constant, and
their five tests. The moved code is byte-identical apart from new module
docs; main.rs keeps its clap definitions, its dispatch match, the two shared
exit helpers, and run_serve. Doc pointers that named main.rs (cli/mod.rs's
module docs, main.rs's own header, DOC-51's implementation table) now name
cli::legacy_run_task.

main.rs: 498 -> 335 SLOC (165 of headroom under the cap).
Behaviour-preserving: no test expectation was modified, and the binary test
target still runs the same 11 tests.

Test: cargo test -p trusty-code (1590 lib + 11 bin + 58 integration, all
pass), cargo clippy -p trusty-code --all-targets -- -D warnings, cargo fmt
--check, bash scripts/check_line_cap.sh, bash scripts/check_test_pointers.sh
— all clean.

Refs #4434
